+++ package.json
@@ -41,6 +41,7 @@
     ]
   },
   "devDependencies": {
+    "react-router-dom": "^6.28.0",
     "tailwindcss": "^3.4.14"
   }
 }
