+++ src/config/navbar.config.js
@@ -1,0 +1,7 @@
+const NAV_LINKS = [
+  {title: "Classes", url: "/classes"},
+  {title: "Chats", url: "/chats"},
+  {title: "Profile", url: "/profile"}
+];
+
+export default NAV_LINKS;
